Controlling progress in other class
